backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
Pillow>=10.0
python-decouple>=3.8
django-filter>=23.0